feat:refatoração do certificado
