+++ site Scrum do Zero/src/services/certificado.service.js
@@ -1,0 +1,11 @@
+const {
+  findCertificadoByHash,
+} = require("../repositories/certificados.repository");
+
+async function buscarCertificadoPorHash(certificadoHash) {
+  return findCertificadoByHash(certificadoHash);
+}
+
+module.exports = {
+  buscarCertificadoPorHash,
+};
